+++ packages/backend-function/src/factory.ts
@@ -11,11 +11,11 @@ import {
   ResourceNameValidator,
   ResourceProvider,
   SsmEnvironmentEntry,
+  StackProvider,
 } from '@aws-amplify/plugin-types';
 import { Construct } from 'constructs';
 import { NodejsFunction, OutputFormat } from 'aws-cdk-lib/aws-lambda-nodejs';
 import * as path from 'path';
-import { getCallerDirectory } from './get_caller_directory.js';
 import { Duration, Stack, Tags } from 'aws-cdk-lib';
 import { CfnFunction, Runtime } from 'aws-cdk-lib/aws-lambda';
 import { createRequire } from 'module';
@@ -30,7 +30,11 @@ import {
 import { FunctionEnvironmentTypeGenerator } from './function_env_type_generator.js';
 import { AttributionMetadataStorage } from '@aws-amplify/backend-output-storage';
 import { fileURLToPath } from 'node:url';
-import { AmplifyUserError, TagName } from '@aws-amplify/platform-core';
+import {
+  AmplifyUserError,
+  CallerDirectoryExtractor,
+  TagName,
+} from '@aws-amplify/platform-core';
 import { convertFunctionSchedulesToRuleSchedules } from './schedule_parser.js';
 import * as targets from 'aws-cdk-lib/aws-events-targets';
 import { Rule } from 'aws-cdk-lib/aws-events';
@@ -61,7 +65,8 @@ export const defineFunction = (
 ): ConstructFactory<
   ResourceProvider<FunctionResources> &
     ResourceAccessAcceptorFactory &
-    AddEnvironmentFactory
+    AddEnvironmentFactory &
+    StackProvider
 > => new FunctionFactory(props, new Error().stack);
 
 export type FunctionProps = {
@@ -118,7 +123,7 @@ export type FunctionProps = {
    * @example
    * schedule: "every week"
    * @example
-   * schedule: "0 9 * * 2" // every Monday at 9am
+   * schedule: "0 9 ? * 2 *" // every Monday at 9am
    */
   schedule?: FunctionSchedule | FunctionSchedule[];
 };
@@ -180,13 +185,18 @@ class FunctionFactory implements ConstructFactory<AmplifyFunction> {
     }
 
     // Otherwise, use the directory name where the function is defined
-    return path.basename(getCallerDirectory(this.callerStack));
+    return path.basename(
+      new CallerDirectoryExtractor(this.callerStack).extract()
+    );
   };
 
   private resolveEntry = () => {
     // if entry is not set, default to handler.ts
     if (!this.props.entry) {
-      return path.join(getCallerDirectory(this.callerStack), 'handler.ts');
+      return path.join(
+        new CallerDirectoryExtractor(this.callerStack).extract(),
+        'handler.ts'
+      );
     }
 
     // if entry is absolute use that
@@ -195,7 +205,10 @@ class FunctionFactory implements ConstructFactory<AmplifyFunction> {
     }
 
     // if entry is relative, compute with respect to the caller directory
-    return path.join(getCallerDirectory(this.callerStack), this.props.entry);
+    return path.join(
+      new CallerDirectoryExtractor(this.callerStack).extract(),
+      this.props.entry
+    );
   };
 
   private resolveTimeout = () => {
@@ -297,6 +310,7 @@ class AmplifyFunction
     AddEnvironmentFactory
 {
   readonly resources: FunctionResources;
+  readonly stack: Stack;
   private readonly functionEnvironmentTranslator: FunctionEnvironmentTranslator;
   constructor(
     scope: Construct,
@@ -306,6 +320,8 @@ class AmplifyFunction
     outputStorageStrategy: BackendOutputStorageStrategy<FunctionOutput>
   ) {
     super(scope, id);
+
+    this.stack = Stack.of(scope);
 
     const runtime = nodeVersionMap[props.runtime];
 
@@ -362,11 +378,22 @@ class AmplifyFunction
         },
       });
     } catch (error) {
+      // If the error is from ES Bundler which is executed as a child process by CDK,
+      // then the error from CDK contains the command that was executed along with the exit status.
+      // Wrapping it here  would cause the cdk_deployer to re-throw this wrapped exception
+      // instead of scraping the stderr for actual ESBuild error.
+      if (
+        error instanceof Error &&
+        error.message.match(/Failed to bundle asset.*exited with status/)
+      ) {
+        throw error;
+      }
       throw new AmplifyUserError(
         'NodeJSFunctionConstructInitializationError',
         {
           message: 'Failed to instantiate nodejs function construct',
-          resolution: 'See the underlying error message for more details.',
+          resolution:
+            'See the underlying error message for more details. Use `--debug` for additional debugging information.',
         },
         error as Error
       );