+++ packages/form-generator/src/local_codegen_graphql_form_generator.ts
@@ -222,11 +222,13 @@ export class LocalGraphqlFormGenerator implements GraphqlFormGenerator {
             ([key]) => key.toLowerCase() === model.toLowerCase()
           );
           if (!entry) {
+            // eslint-disable-next-line amplify-backend-rules/prefer-amplify-errors
             throw new Error(`Could not find specified model ${model}`);
           }
           prev.push(entry);
           return prev;
         }
+        // eslint-disable-next-line amplify-backend-rules/prefer-amplify-errors
         throw new Error(`Could not find specified model ${model}`);
       },
       []